refactor(contracts): remove redundant re-exports from contracts.py
Remove backward-compatibility re-exports (ExecutionLaunchResult,
ExecutionRequest, WorktreeRequirement) from vibe3.execution.contracts.
All consumers now import directly from vibe3.models.

Changes:
- Update 10 test files to import from vibe3.models
- Update 1 doc file to use correct import path
- Delete test_contract_compat.py (no longer needed)
- Remove re-export block from contracts.py (lines 6-11)

The file retains production symbols:
- AsyncLauncherProtocol (Protocol class)
- _StartAsyncFactory (type alias)

All tests pass (1016 tests in affected modules).
Type checking passes (413 source files).

Ref: #2507
