+++ tests/vibe3/domain/handlers/test_dispatch_not_launched.py
@@ -6,7 +6,7 @@ Verifies that when coordinator does not launch, handler logs warning without err
 from unittest.mock import MagicMock, patch
 
 from vibe3.domain.events.flow_lifecycle import PlannerDispatchIntent
-from vibe3.execution.contracts import ExecutionLaunchResult, ExecutionRequest
+from vibe3.models import ExecutionLaunchResult, ExecutionRequest
 
 
 def _make_mock_request(
