+++ tests/vibe3/domain/handlers/test_governance_scan.py
@@ -3,7 +3,7 @@
 from unittest.mock import MagicMock, patch
 
 from vibe3.domain.events.governance import GovernanceScanStarted
-from vibe3.execution.contracts import ExecutionLaunchResult
+from vibe3.models import ExecutionLaunchResult
 
 
 class TestGovernanceScanHandler:
