+++ tests/vibe3/execution/test_coordinator_worktree.py
@@ -12,9 +12,9 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
-from vibe3.execution.contracts import ExecutionRequest
 from vibe3.execution.coordinator import ExecutionCoordinator
 from vibe3.execution.role_contracts import WorktreeRequirement
+from vibe3.models import ExecutionRequest
 
 
 @pytest.fixture
