+++ tests/vibe3/services/test_shared_errors_exception.py
@@ -2,7 +2,7 @@
 
 from unittest.mock import MagicMock, patch
 
-from vibe3.execution.contracts import ExecutionLaunchResult
+from vibe3.models import ExecutionLaunchResult
 from vibe3.services.shared.errors import record_dispatch_failure_if_unexpected
